Look up the thumbnail in DisplayImage, not DerivedData
