+++ common/display_model/src/main/scala/weco/catalogue/display_model/image/DisplayImage.scala
@@ -3,6 +3,7 @@ package weco.catalogue.display_model.image
 import io.circe.generic.extras.JsonKey
 import weco.catalogue.display_model.locations.DisplayDigitalLocation
 import weco.catalogue.internal_model.image.{Image, ImageState}
+import weco.catalogue.internal_model.locations.{DigitalLocation, LocationType}
 
 case class DisplayImage(
   id: String,
@@ -13,11 +14,19 @@ case class DisplayImage(
 )
 
 object DisplayImage {
+  private def thumbnail(image: Image[ImageState.Indexed]): DigitalLocation =
+    image.locations
+      .find(_.locationType == LocationType.IIIFImageAPI)
+      .getOrElse(
+        // This should never happen
+        throw new RuntimeException(
+          s"No iiif-image (thumbnail) location found on image ${image.sourceIdentifier}")
+      )
 
   def apply(image: Image[ImageState.Indexed]): DisplayImage =
     new DisplayImage(
       id = image.id,
-      thumbnail = DisplayDigitalLocation(image.state.derivedData.thumbnail),
+      thumbnail = DisplayDigitalLocation(thumbnail(image)),
       locations = image.locations.map(DisplayDigitalLocation(_)),
       source = DisplayImageSource(image.source)
     )
